extracted refusal utils + new xstest dataset pointer
