+++ llm_behavior_eval/evaluation_utils/custom_dataset.py
@@ -2,7 +2,7 @@ import logging
 from copy import copy
 from functools import partial
 from pathlib import Path
-from typing import Any, cast
+from typing import cast
 
 import torch
 from datasets import Dataset, DatasetDict, load_dataset
@@ -11,16 +11,13 @@ from transformers.tokenization_utils_base import PreTrainedTokenizerBase
 from .dataset_config import PreprocessConfig
 from .enums import DatasetType
 from .prompts import SYSTEM_PROMPT_DICT
+from .refusal_utils import (
+    UNSAFE_REFUSAL_LABEL,
+    is_refusal_dataset,
+    normalize_refusal_dataset,
+    normalize_refusal_label,
+)
 from .util_functions import is_model_multimodal, safe_apply_chat_template
-
-REFUSAL_PLACEHOLDER_ANSWER = "placeholder"
-SAFE_REFUSAL_LABEL = "safe"
-UNSAFE_REFUSAL_LABEL = "unsafe"
-REFUSAL_DATASETS = {"walledai/XSTest", "hirundo-io/or-bench"}
-
-
-def is_refusal_dataset(file_path: Path | str) -> bool:
-    return str(file_path) in REFUSAL_DATASETS
 
 
 def validate_dataset_columns(hf_dataset: Dataset) -> None:
@@ -36,59 +33,6 @@ def validate_dataset_columns(hf_dataset: Dataset) -> None:
         raise ValueError(
             f"Dataset is missing required columns: {missing}; found {hf_dataset.column_names}"
         )
-
-
-def normalize_refusal_label(raw_label: Any) -> str:
-    label = str(raw_label).strip().lower()
-    if label not in {SAFE_REFUSAL_LABEL, UNSAFE_REFUSAL_LABEL}:
-        raise ValueError(
-            "Refusal dataset label must be "
-            f"{SAFE_REFUSAL_LABEL!r} or {UNSAFE_REFUSAL_LABEL!r}, got {raw_label!r}"
-        )
-    return label
-
-
-def normalize_refusal_dataset(hf_dataset: Dataset) -> Dataset:
-    """Normalize refusal datasets to the free-text schema expected downstream."""
-
-    def _normalize_batch(examples_batch: dict[str, list[Any]]) -> dict[str, list[str]]:
-        prompts = examples_batch.get("question")
-        if prompts is None:
-            prompts = examples_batch.get("prompt")
-        if prompts is None:
-            raise ValueError(
-                "Refusal dataset must contain either a 'question' or 'prompt' column"
-            )
-        labels = examples_batch.get("label")
-        if labels is None:
-            raise ValueError("Refusal dataset must contain a 'label' column")
-
-        normalized_labels = [normalize_refusal_label(label) for label in labels]
-        answers = examples_batch.get("answer")
-        if answers is None:
-            # Free-text preprocessing expects an "answer" column, but our refusal
-            # datasets currently only provide prompt + safe/unsafe label.
-            answers = [REFUSAL_PLACEHOLDER_ANSWER for _ in prompts]
-        else:
-            answers = [
-                answer if str(answer).strip() else REFUSAL_PLACEHOLDER_ANSWER
-                for answer in answers
-            ]
-
-        return {
-            "question": [str(prompt) for prompt in prompts],
-            "answer": [str(answer) for answer in answers],
-            "label": normalized_labels,
-        }
-
-    return cast(
-        "Dataset",
-        hf_dataset.map(
-            _normalize_batch,
-            batched=True,
-            remove_columns=hf_dataset.column_names,
-        ),
-    )
 
 
 def free_text_preprocess_function(
